# Task3/server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API для моніторингу робочих умов",
    "description": "API для роботи з даними користувачів, локаціями, сенсорами, сповіщеннями та повідомленнями.",
    "version": "1.0.0"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Отримання списку всіх користувачів",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Список користувачів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user_id": { "type": "string" },
                      "first_name": { "type": "string" },
                      "last_name": { "type": "string" },
                      "email": { "type": "string" },
                      "role": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Реєстрація нового користувача",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "role": {
                    "type": "string",
                    "enum": ["Employee", "Administrator"]
                  },
                  "phone": { "type": "string" },
                  "location_id": { "type": "string" }
                },
                "required": ["first_name", "last_name", "email", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Користувача успішно зареєстровано",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "user_id": { "type": "string" },
                        "first_name": { "type": "string" },
                        "last_name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Помилка при реєстрації користувача"
          },
          "409": {
            "description": "Користувач з таким email вже існує"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Отримання даних користувача за ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID користувача",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Дані користувача",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": { "type": "string" },
                    "first_name": { "type": "string" },
                    "last_name": { "type": "string" },
                    "email": { "type": "string" },
                    "role": { "type": "string" },
                    "phone": { "type": "string" },
                    "location_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Користувача не знайдено" }
        }
      },
      "put": {
        "summary": "Редагування даних користувача",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID користувача, якого потрібно редагувати",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "email": { "type": "string" },
                  "role": { "type": "string" },
                  "phone": { "type": "string" },
                  "location_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Дані користувача успішно оновлено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "user_id": { "type": "string" },
                        "first_name": { "type": "string" },
                        "last_name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" },
                        "phone": { "type": "string" },
                        "location_id": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Користувача не знайдено" },
          "400": { "description": "Помилка при оновленні користувача" }
        }
      },
      "delete": {
        "summary": "Видалення користувача",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID користувача, якого потрібно видалити",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": { "description": "Користувача успішно видалено" },
          "404": { "description": "Користувача не знайдено" },
          "400": { "description": "Помилка при видаленні користувача" }
        }
      }
    },
...
    "/sensors": {
      "get": {
        "summary": "Отримання всіх сенсорів",
        "operationId": "getSensors",
        "responses": {
          "200": {
            "description": "Список всіх сенсорів",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "sensor_id": { "type": "string" },
                      "location_id": { "type": "string" },
                      "type": { "type": "string" },
                      "value": { "type": "number" },
                      "timestamp": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Додавання нового сенсора",
        "operationId": "addSensor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location_id": { "type": "string" },
                  "type": { "type": "string" },
                  "value": { "type": "number" }
                },
                "required": ["location_id", "type", "value"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Сенсор успішно додано",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sensor_id": { "type": "string" },
                    "location_id": { "type": "string" },
                    "type": { "type": "string" },
                    "value": { "type": "number" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Помилка при додаванні сенсора"
          }
        }
      }
    },
    "/sensors/{id}": {
      "get": {
        "summary": "Отримання сенсора за ID",
        "operationId": "getSensorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID сенсора",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Сенсор знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sensor_id": { "type": "string" },
                    "location_id": { "type": "string" },
                    "type": { "type": "string" },
                    "value": { "type": "number" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Сенсор не знайдено"
          }
        }
      },
      "put": {
        "summary": "Редагування сенсора за ID",
        "operationId": "updateSensor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID сенсора",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location_id": { "type": "string" },
                  "type": { "type": "string" },
                  "value": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Сенсор успішно оновлено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sensor_id": { "type": "string" },
                    "location_id": { "type": "string" },
                    "type": { "type": "string" },
                    "value": { "type": "number" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Сенсор не знайдено"
          },
          "400": {
            "description": "Помилка при оновленні сенсора"
          }
        }
      },
      "delete": {
        "summary": "Видалення сенсора за ID",
        "operationId": "deleteSensor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID сенсора",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Сенсор успішно видалено"
          },
          "404": {
            "description": "Сенсор не знайдено"
          },
          "400": {
            "description": "Помилка при видаленні сенсора"
          }
        }
      }
    },
...
  }
}